requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
tqdm==4.61.0
urllib3==1.26.5
Werkzeug==2.0.1
celery==5.1.2
redis==3.5.3
flower==1.0.0
sqlmodel==0.0.4
psycopg2==2.9.1
